add new types

+++ Spravy.Domain/Interfaces/IToDoSubItem.cs
@@ -7,7 +7,7 @@ public interface IToDoSubItem
 {
     public Guid Id { get; }
     public string Name { get; }
-    public ulong OrderIndex { get; }
+    public uint OrderIndex { get; }
     public ToDoItemStatus Status { get; }
     public string Description { get; }
     bool IsCurrent { get; }
